tests/wpt/web-platform-tests/css/css-scroll-snap/parsing/scroll-snap-type-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Scroll Snap Test: scroll-snap-type with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-scroll-snap-1/#propdef-scroll-snap-type">
<meta name="assert" content="scroll-snap-type supports only the grammar 'none | [ x | y | block | inline | both ] [ mandatory | proximity ]?'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("scroll-snap-type", "auto");

test_invalid_value("scroll-snap-type", "x y");
test_invalid_value("scroll-snap-type", "block mandatory inline");

test_invalid_value("scroll-snap-type", "mandatory");
test_invalid_value("scroll-snap-type", "proximity");
test_invalid_value("scroll-snap-type", "mandatory inline");
test_invalid_value("scroll-snap-type", "proximity both");
test_invalid_value("scroll-snap-type", "mandatory x");
test_invalid_value("scroll-snap-type", "proximity y");
test_invalid_value("scroll-snap-type", "mandatory block");
test_invalid_value("scroll-snap-type", "proximity mandatory");
</script>
</body>
</html>
